Rename and reorder the modules
